requirements.txt
Django>=5.0
django-tenants>=3.6.0
djangorestframework>=3.14.0
django-allauth>=0.59.0
PyJWT>=2.8.0
psycopg2-binary>=2.9.9
